refactor: replace useGlobalMetrics with useDashboardData in dashboard component
- Removed useGlobalMetrics hook and its associated state management.
- Integrated useDashboardData for fetching dashboard-related data.
- Updated error handling to reflect changes in data fetching.
- Adjusted data structure passed to DataView component for improved clarity.

Closes #274
