+++ app/dashboard/page.tsx
@@ -3,9 +3,7 @@
 import { useEffect, useState } from "react";
 import { DashboardData, dashboardSections } from "@/ui/dataview/datasections/dashboard";
 import DataView from "@/ui/common/data-view-columns";
-import { useChain } from "@cosmos-kit/react";
-import { useVeranaChain } from "@/hooks/useVeranaChain";
-import { useGlobalMetrics } from "@/hooks/useGlobalMetrics";
+import { useDashboardData } from "@/hooks/useDashboardData";
 import { useNotification } from "@/ui/common/notification-provider";
 import TitleAndButton from "@/ui/common/title-and-button";
 import { resolveTranslatable } from "@/ui/dataview/types";
@@ -19,61 +17,29 @@ import DashboardFooter from "@/ui/common/dashboard-footer";
 
 export default function Page() {
 
-  const veranaChain = useVeranaChain();
-  const { getStargateClient, status, isWalletConnected, address, wallet } = useChain(veranaChain.chain_name);
-  const [blockHeight, setBlockHeight] = useState<string>("");
-  const { metrics, errorMetrics } = useGlobalMetrics();
+  const { dashboardData, errorDashboardData, isWalletConnected } = useDashboardData();
   const [errorNotified, setErrorNotified] = useState(false);
   const { notify } = useNotification();
 
-  const data: DashboardData = {
-    chainName: isWalletConnected
-      ? `${veranaChain.chain_name} (${veranaChain.chain_id})`
-      : null,
-    blockHeight,
-    status,
-    isWalletConnected: String(isWalletConnected),
-    address: address ? String(address) : null,
-    walletPrettyName: wallet ? wallet.prettyName : null,
-    ecosystems: metrics ? Number(metrics.active_trust_registries).toLocaleString() : null,
-    schemas: metrics ? Number(metrics.active_schemas).toLocaleString() : null,
-    totalLockedTrustDeposit: metrics ? `${Number(metrics.weight).toLocaleString()} VNA` : null,
-    issuedCredentials: metrics ? Number(metrics.issued).toLocaleString() : null,
-    verifiedCredentials: metrics ? Number(metrics.verified).toLocaleString() : null,
-  };
-
   useEffect(() => {
-    const fetchHeight = async () => {
-      if (getStargateClient && isWalletConnected) {
-        const client = await getStargateClient();
-        const block = await client.getBlock();
-        setBlockHeight(String(block.header.height.toLocaleString(undefined)));
-      }
-    };
-    fetchHeight();
-    const interval = setInterval(fetchHeight, 5000);
-    return () => clearInterval(interval);
-  }, [getStargateClient, isWalletConnected]);
-
-  useEffect(() => {
-    if (errorMetrics && !errorNotified) {
+    if (errorDashboardData && !errorNotified) {
       (async () => {
-        await notify(errorMetrics, 'error', resolveTranslatable({key: "error.fetch.metrics.title"}, translate));
+        await notify(errorDashboardData, 'error', resolveTranslatable({key: "error.fetch.metrics.title"}, translate));
         setErrorNotified(true);
       })();
     }
-  }, [errorMetrics, errorNotified]);
+  }, [errorDashboardData, errorNotified]);
 
   return (
     <>
-      <TitleAndButton 
+      <TitleAndButton
         title={resolveTranslatable({key: "dashboard.title"}, translate)?? "Dashboard"}
         description={[resolveTranslatable({key: "dashboard.desc"}, translate)??""]}
       />
 
       <DataView<DashboardData>
           sectionsI18n={dashboardSections}
-          data={data}
+          data={dashboardData}
           id=""
       />
 
@@ -87,7 +53,7 @@ export default function Page() {
                   <h2 className="text-2xl font-bold text-white mb-2">{resolveTranslatable({key: 'notconnected.dashboard.title'}, translate)}</h2>
                   <p className="text-primary-100 text-lg mb-4">{resolveTranslatable({key: 'notconnected.dashboard.msg'}, translate)}</p>
               </div>
-              <div className="flex-shrink-0">                  
+              <div className="flex-shrink-0">
                   <div className="inline-flex items-center px-8 py-4 bg-white text-primary-700 text-lg font-semibold rounded-xl hover:bg-gray-50 transition-all duration-200 shadow-lg hover:shadow-xl transform hover:-translate-y-1 focus:outline-none focus:ring-2 focus:ring-white focus:ring-offset-2 focus:ring-offset-primary-600">
                     <FontAwesomeIcon icon={faWallet} />
                     <Wallet isNavBar={false}/>
